requirements.txt
fastapi
uvicorn[standard]
pydantic
joblib
pandas
python-dotenv
scikit-learn==1.4.2
lightgbm
pytest
